requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
tensorflow==2.15.0
pyspellchecker==0.8.4